+++ scripts/verify-lockfile.ts
@@ -4,7 +4,7 @@ import { x } from "tinyexec";
 const { stdout } = await x("git", ["diff", "HEAD@{1}", "--stat", "--", "./pnpm-lock.yaml"]);
 
 if (stdout.split("\n").length > 0) {
-  console.log("");
+  console.info("");
   console.info(ansis.yellow("Detected changes in pnpm-lock.yaml!"));
   console.info(
     ansis.yellowBright(